app/requirements.txt
numpy
tensorflow
Pillow
streamlit
google-generativeai
python-dotenv
protobuf